pyproject.toml
[tool.poetry]
name = "student-centric-teacher-evaluation-system"
version = "0.1.0"
description = ""
authors = ["Elyson"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.115.0"
uvicorn = "0.31.0"
sqlalchemy = "2.0.35"
alembic = "1.13.3"
pydantic = "2.9.2"
python-dotenv = "1.0.1"
sqlalchemy-utils = "0.41.2"
psycopg2-binary = "2.9.9"
fastapi-pagination = "0.12.29"
python-jose = "3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
cryptography = "43.0.3"
python-multipart = "0.0.12"


[tool.poetry.group.dev.dependencies]
pytest = "8.3.3"
ruff = "0.6.8"
pre-commit = "4.0.0"
httpx = "^0.27.2"
pytest-cov = "5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py312"
exclude = ["alembic"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint]
# Add the `line-too-long` rule to the enforced rule set.
extend-select = ["E501","D"]
select = ["E4", "E7", "E9", "F"]
ignore = []
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
skip-magic-trailing-comma = false
line-ending = "auto"